refactor: collection 패키지 삭제
